sdk: tests: refactor: move sleep to tests/utils.ts
