+++ raiden-ts/tests/integration/epics/path.spec.ts
@@ -10,15 +10,7 @@ import {
   token,
   tokenNetwork,
 } from '../fixtures';
-import {
-  fetch,
-  makeLog,
-  makeRaiden,
-  makeRaidens,
-  providersEmit,
-  sleep,
-  waitBlock,
-} from '../mocks';
+import { fetch, makeLog, makeRaiden, makeRaidens, providersEmit, waitBlock } from '../mocks';
 
 import { defaultAbiCoder } from '@ethersproject/abi';
 import { BigNumber } from '@ethersproject/bignumber';
@@ -38,7 +30,7 @@ import { ErrorCodes } from '@/utils/error';
 import type { Address, Int, Signature, UInt } from '@/utils/types';
 import { Signed } from '@/utils/types';
 
-import { makeAddress } from '../../utils';
+import { makeAddress, sleep } from '../../utils';
 import type { MockedRaiden } from '../mocks';
 
 const pfsAddress = makeAddress();
