+++ raiden-ts/tests/integration/epics/send.spec.ts
@@ -9,7 +9,7 @@ import {
   secrethash,
   tokenNetwork,
 } from '../fixtures';
-import { makeLog, makeRaiden, makeRaidens, providersEmit, sleep, waitBlock } from '../mocks';
+import { makeLog, makeRaiden, makeRaidens, providersEmit, waitBlock } from '../mocks';
 
 import { BigNumber } from '@ethersproject/bignumber';
 import { Two, Zero } from '@ethersproject/constants';
@@ -37,7 +37,7 @@ import { getSecrethash, makePaymentId, makeSecret, transferKey } from '@/transfe
 import { isResponseOf } from '@/utils/actions';
 import type { Int, UInt } from '@/utils/types';
 
-import { makeHash } from '../../utils';
+import { makeHash, sleep } from '../../utils';
 import type { MockedRaiden } from '../mocks';
 
 const direction = Direction.SENT;
