+++ raiden-ts/tests/integration/epics/withdraw.spec.ts
@@ -8,7 +8,7 @@ import {
   id,
   tokenNetwork,
 } from '../fixtures';
-import { makeRaidens, makeTransaction, sleep, waitBlock } from '../mocks';
+import { makeRaidens, makeTransaction, waitBlock } from '../mocks';
 
 import { BigNumber } from '@ethersproject/bignumber';
 import { Zero } from '@ethersproject/constants';
@@ -26,6 +26,7 @@ import { makeMessageId } from '@/transfers/utils';
 import { ErrorCodes } from '@/utils/error';
 import type { Hash, UInt } from '@/utils/types';
 
+import { sleep } from '../../utils';
 import type { MockedRaiden } from '../mocks';
 
 describe('withdraw receive request', () => {
